requirement.txt
ChatterBot-0.8.4
simple-websocket-server